pyproject.toml
[tool.poetry]
name = "course-project"
version = "0.1.0"
description = ""
authors = ["islam <islamdomka2005@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"


[tool.poetry.group.dev.dependencies]
pandas = "^2.2.3"
logging = "^0.4.9.6"
pytest = "^8.3.3"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
pytest-cov = "^6.0.0"
openpyxl = "^3.1.5"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.1"
isort = "^5.13.2"
black = "^24.10.0"
mypy = "^1.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
